Created a GUI and added/modified functions
I have created a GUI and added functions for the buttons and listbox to work. I have also modified the original script so that the user is able to search through any folder they wish  and they can select and copy the files that have been updated within the last 24 hours to the folder they desire.