updated the code to use a database, corrected functions
I have updated the code to use sqlite3 as a database to insert timestamps and pull them to display in a label for users. 
I have updated the code to side step future errors the user might encounter if they cancel the filedialog.askdirectory(). 
I have created a function that will pull the timestamp only if this program has ran previously.